// src/types/AuthorInfo.tsx
export type AuthorInfo = {
	first_name: string;
	last_name: string;
	date_of_birth: string;
	date_of_death: string;
	biography: string;
	nationality: string;
	contact_email: string;
	website: string;
	awards_and_honors: string;
};

// export type AuthorInfo = {
// 	first_name: string;
// 	last_name: string;
// 	date_of_birth?: string;
// 	date_of_death?: string;
// 	biography?: string;
// 	nationality?: string;
// 	contact_email: string;
// 	website?: string;
// 	awards_and_honors?: string;
// };